Removed unused code, cleaned it up and added a log category
All files that I changed are marked with "//TO DO". I cleaned almost all the code files in the whole project, it was a lot of unused commented out code but also functions and variables not used. It's much easier to read and navigate the code now. I also added a custom Log category for easier debugging of ability states and so on.
However, If I were to *really* refactor the code, I would probably have to rewrite whole systems(which would take a lot of time, hence me not doing so).  I think the code quality is okay for this short project now, I also think that removing unused stuff might've actually removed some bugs. I can't really find any bugs as of right now.